Renamed a parameter name.
Made the default clear color the same color as is used
in the drop down combo.

Remove commented out C-structs from the python code.
I will not attempt to create ctypes bindings for all
structs.  I tried that, it was unfruitful.

For structs which are created by the caller, I will
make bindings for them.  For structs created by the C
code, no bindings shall be made, they will just
be pointers.

For anything which can't follow that pattern, C-glue
code should be made.